Update README.md
Using onnxruntime-gpu 1.10 (last one), the following error will occur:

raise ValueError("This ORT build has {} enabled. ".format(available_providers) +
ValueError: This ORT build has ['TensorrtExecutionProvider', 'CUDAExecutionProvider', 'CPUExecutionProvider'] enabled. Since ORT 1.9, you are required to explicitly set the providers parameter when instantiating InferenceSession. For example, onnxruntime.InferenceSession(..., providers=['TensorrtExecutionProvider', 'CUDAExecutionProvider', 'CPUExecutionProvider'], ...)

To solve this is enough to pass the providers to the prepare function. A better solution may be to pass a default value to the onnx session, but for the moment this should do the trick.